Move sua logic to core
